requirements.txt
# requirements.txt
streamlit
pandas
pymongo
plotly.express